% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_range_bins.R
\name{calculate_range_bins}
\alias{calculate_range_bins}
\title{Calculate bins of 3, 5, 7 of proportional rate for RADI dataset}
\usage{
calculate_range_bins(df)
}
\description{
Calculate bins of 3, 5, 7 of proportional rate for RADI dataset
}
