% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_coefs.R
\name{make_coefs}
\alias{make_coefs}
\title{Create coefficients dataframes to use in in-line R reporting.}
\usage{
make_coefs(model, mod_name)
}
\description{
Create coefficients dataframes to use in in-line R reporting.
}
