% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_tg_wav.R
\name{open_tg_wav}
\alias{open_tg_wav}
\title{Open a .wav and corresponding .TextGrid in system viewer (e.g., Praat)}
\usage{
open_tg_wav(filename, wPath = wavPath, tPath = tgPath)
}
\description{
Open a .wav and corresponding .TextGrid in system viewer (e.g., Praat)
}
