% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thear.R
\name{radi_filename_prep}
\alias{radi_filename_prep}
\title{Split filenames from RADI into meaningful units}
\usage{
radi_filename_prep(df)
}
\arguments{
\item{df}{Dataframe containing a column filename to be split}
}
\description{
This is designed to be used with the RADI data
}
