% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thear.R
\name{remove_duplicates}
\alias{remove_duplicates}
\title{Remove duplicate files from visual analog scale results}
\usage{
remove_duplicates(file)
}
\description{
Remove duplicate files from visual analog scale results
}
