% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_habitual_mean.R
\name{calculate_habitual_mean}
\alias{calculate_habitual_mean}
\title{Calculate habitual mean for RADI dataset}
\usage{
calculate_habitual_mean(df)
}
\arguments{
\item{df}{Data frame with columns "participant", "utterance_duration", "rate". Habitual "rate" must *must* be called "H1"}
}
\description{
Calculate habitual mean for RADI dataset
}
